fix liftspeed bug with resizable dash switch
This makes the switch inherit the liftspeed of the platform it's attached to, which alleviates a really complicated and annoying bug where the player actually gets liftspeed from the attached switch rather than the main block, even if the player isn't actually touching the switch. This doesn't actually _fix_ that bug, it just makes it not matter. I hope.